workspace-wide T7 + clippy + rustdoc gate-tightening
Final sweep on top of PR #163 surfaced 1 T7-attributable HIGH gap and 21
pre-existing latent clippy errors + 2 rustdoc broken-link errors across the
workspace. Per rule 1 (all code is production) all are fixed:

T7-attributable (HIGH):
- .github/workflows/ci-pool.yml: previously only ran pool-proxy-smoke.sh
  manually. Now also runs the two new T7 live smokes
  (pool-extended-query-pipeline-live-smoke.sh + through-pool variant), plus
  workspace-tight gates: cargo clippy -p ai_blaise_citus_pool
  -p ai_blaise_citus_pool_wire --all-targets -- -D warnings and
  RUSTDOCFLAGS=-D rustdoc::broken-intra-doc-links cargo doc ... --no-deps.
  This makes the T7 production-ready claim CI-enforced, not local-only.

Pre-existing clippy errors fixed workspace-wide (cargo clippy --fix
auto-resolved most; 4 manual):
- tools/citusctl/src/lib.rs:2204-2208 (5x unnecessary_cast u32->u32)
- tools/citus-lsp/src/{lib,main}.rs (collapsible if, manual char compare,
  while-let-loop)
- sidecar/auth/src/lib.rs (needless_as_bytes, 4x manual_div_ceil,
  needless_range_loop on SHA-1 mix loop via targeted allow)
- sidecar/backup/src/{lib,scheduler}.rs (manual_is_multiple_of x4,
  let-else -> ? operator on json_u32)
- sidecar/postgrest/src/lib.rs (1 error)
- sidecar/raft/src/{runtime,main}.rs (3 errors)
- sidecar/realtime/src/{live,ws}.rs (manual_div_ceil, doc_lazy_continuation
  module-level allow), tests/realtime_e2e.rs (field_reassign_with_default,
  useless_format)
- sidecar/edge_functions/src/lib.rs (too_many_arguments via targeted allow
  with EF6 alpha follow-up note)

Rustdoc broken-link errors fixed:
- sidecar/vectorizer/src/runtime/worker.rs:430 (shutdown_handle ref)
- sidecar/cdc/src/dlq.rs:5 (DeliveryRetryPolicy::max_attempts ref)

VM-verified on experiment-playground-plus-2 after this commit:
- cargo clippy --workspace --all-targets -- -D warnings: clean
- RUSTDOCFLAGS=-D rustdoc::broken-intra-doc-links cargo doc --workspace --no-deps: clean
- cargo test --workspace --lib: all 31 workspace members pass
- All 6 release gates pass: production_ready=276 alpha_headings=0
- pool-extended-query-pipeline-live-smoke + through-pool + pool-proxy-smoke: all pass
- Pool ext-query counters: Parse=3 Bind=4 Describe=3 Execute=4 Sync=4 Flush=1 decode_errors=0

+++ tools/citusctl/src/lib.rs
@@ -2201,11 +2201,11 @@ fn utc_timestamp_epoch_seconds(value: &str) -> Result<i64, CitusCtlError> {
         return Err(CitusCtlError::InvalidTimestamp);
     }
     let year = parse_fixed_digits(&value[0..4])? as i32;
-    let month = parse_fixed_digits(&value[5..7])? as u32;
-    let day = parse_fixed_digits(&value[8..10])? as u32;
-    let hour = parse_fixed_digits(&value[11..13])? as u32;
-    let minute = parse_fixed_digits(&value[14..16])? as u32;
-    let second = parse_fixed_digits(&value[17..19])? as u32;
+    let month = parse_fixed_digits(&value[5..7])?;
+    let day = parse_fixed_digits(&value[8..10])?;
+    let hour = parse_fixed_digits(&value[11..13])?;
+    let minute = parse_fixed_digits(&value[14..16])?;
+    let second = parse_fixed_digits(&value[17..19])?;
 
     if !(1..=12).contains(&month) || hour > 23 || minute > 59 || second > 59 {
         return Err(CitusCtlError::InvalidTimestamp);
